cleanup config
