+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/PureConfigReaderWriter.scala
@@ -21,6 +21,8 @@ import com.daml.platform.configuration.{
   CommandConfiguration,
   IndexServiceConfig,
   InitialLedgerConfiguration,
+  TransactionsFlatStreamReaderConfig,
+  TransactionsTreeStreamReaderConfig,
 }
 import com.daml.platform.indexer.ha.HaConfig
 import com.daml.platform.indexer.{IndexerConfig, IndexerStartupMode, PackageMetadataViewConfig}
@@ -333,6 +335,14 @@ class PureConfigReaderWriter(secure: Boolean = true) {
 
   implicit val indexServiceConfigHint =
     ProductHint[IndexServiceConfig](allowUnknownKeys = false)
+
+  implicit val transactionsTreeStreamReaderConfigConfigConvert
+      : ConfigConvert[TransactionsTreeStreamReaderConfig] =
+    deriveConvert[TransactionsTreeStreamReaderConfig]
+
+  implicit val transactionsFlatStreamReaderConfigConfigConvert
+      : ConfigConvert[TransactionsFlatStreamReaderConfig] =
+    deriveConvert[TransactionsFlatStreamReaderConfig]
 
   implicit val indexServiceConfigConvert: ConfigConvert[IndexServiceConfig] =
     deriveConvert[IndexServiceConfig]
