+++ ledger/participant-integration-api/src/main/scala/platform/configuration/IndexServiceConfig.scala
@@ -26,6 +26,13 @@ final case class IndexServiceConfig(
     preparePackageMetadataTimeOutWarning: FiniteDuration =
       IndexServiceConfig.PreparePackageMetadataTimeOutWarning,
     completionsPageSize: Int = IndexServiceConfig.DefaultCompletionsPageSize,
+    transactionsFlatStreamReaderConfig: TransactionsFlatStreamReaderConfig =
+      TransactionsFlatStreamReaderConfig.default,
+    transactionsTreeStreamReaderConfig: TransactionsTreeStreamReaderConfig =
+      TransactionsTreeStreamReaderConfig.default,
+    // TODO etq: Take care what config key names will get exposed in the user-facing participant config
+    globalMaxIdQueries: Int = 20,
+    globalMaxPayloadQueries: Int = 10,
 )
 
 object IndexServiceConfig {
@@ -47,4 +54,43 @@ object IndexServiceConfig {
   val DefaultInMemoryFanOutThreadPoolSize: Int = 16
   val PreparePackageMetadataTimeOutWarning: FiniteDuration = FiniteDuration(1, "second")
   val DefaultCompletionsPageSize: Int = 1000
+}
+
+// TODO etq: Take care what config key names will get exposed in the user-facing participant config
+case class TransactionsFlatStreamReaderConfig(
+    maxIdsPerIdPage: Int = 20000,
+    maxPagesPerIdPagesBuffer: Int = 1,
+    maxWorkingMemoryInBytesForIdPages: Int = 100 * 1024 * 1024,
+    maxPayloadsPerPayloadsPage: Int = 1000, // eventsPageSize
+    // TODO pbatko: Check if these really must be powers of 2
+    // BEGIN
+    maxParallelIdCreateQueries: Int = 4,
+    maxParallelIdConsumingQueries: Int = 4,
+    maxParallelPayloadCreateQueries: Int = 2,
+    maxParallelPayloadConsumingQueries: Int = 2,
+    // END
+    maxParallelPayloadQueries: Int = 2,
+    payloadProcessingParallelism: Int = 8,
+)
+object TransactionsFlatStreamReaderConfig {
+  def default: TransactionsFlatStreamReaderConfig = TransactionsFlatStreamReaderConfig()
+}
+
+// TODO etq: Take care what config key names will get exposed in the user-facing participant config
+case class TransactionsTreeStreamReaderConfig(
+    maxIdsPerIdPage: Int = 20000,
+    maxPagesPerIdPagesBuffer: Int = 1,
+    maxWorkingMemoryInBytesForIdPages: Int = 100 * 1024 * 1024,
+    maxPayloadsPerPayloadsPage: Int = 1000,
+    maxParallelIdCreateQueries: Int = 8,
+    maxParallelIdConsumingQueries: Int = 8,
+    maxParallelIdNonConsumingQueries: Int = 4,
+    maxParallelPayloadCreateQueries: Int = 2,
+    maxParallelPayloadConsumingQueries: Int = 2,
+    maxParallelPayloadNonConsumingQueries: Int = 2,
+    maxParallelPayloadQueries: Int = 2,
+    payloadProcessingParallelism: Int = 8,
+)
+object TransactionsTreeStreamReaderConfig {
+  def default: TransactionsTreeStreamReaderConfig = TransactionsTreeStreamReaderConfig()
 }
